Main

Description

Version : 1.0

Calculates the time between now and the birth of Lydia Pedersén Bohman in seconds.

Documentation

main :: IO ()

Main function. Start of program.

calcTime :: IO ClockTime

Calculates the time difference between now and birth time for Lydia Pedersén Bohman.